refactor(ui): convert languageselect to css modules and reorganize component structure
Migrated the LanguageSelect component from inline styles to CSS modules for better
maintainability. Moved animations from global CSS to component-specific styles.

Key changes:
- Removed animations.css from global theme imports
- Created dedicated module.scss files for components
- Reorganized component file structure
- Moved keyframe animations to component-specific styles
- Improved style encapsulation.
