--- src/assets/css/theme/animations.css
@@ -1,47 +1,0 @@
-@theme static {
- /* Animations */
-
- --animate-language-close: language-close 0.15s ease-in-out;
- --animate-language-open: language-open 0.15s ease-in-out;
-
- /* Keyframes */
- @keyframes fadeIn {
-  0% {
-   opacity: 0;
-  }
-
-  100% {
-   opacity: 1;
-  }
- }
-
- @keyframes fadeOut {
-  0% {
-   opacity: 1;
-  }
-
-  100% {
-   opacity: 0;
-  }
- }
-
- @keyframes language-close {
-  0% {
-   height: 160px;
-  }
-
-  100% {
-   height: 0;
-  }
- }
-
- @keyframes language-open {
-  0% {
-   height: 0;
-  }
-
-  100% {
-   height: 160px;
-  }
- }
-}
