+++ src/assets/css/theme/theme.css
@@ -1,4 +1,3 @@
-@import "./animations.css";
-@import "./colors.css";
-@import "./sizes.css";
-@import "./text.css";
+@import url("./colors.css");
+@import url("./sizes.css");
+@import url("./text.css");
