+++ src/shared/LanguageSelect/index.ts
@@ -1,2 +1,2 @@
 export { type LangItem } from "./model/language.type";
-export { LanguageSelect } from "./ui/LanguageSelect";
+export { LanguageSelect } from "./ui/LanguageSelect/LanguageSelect";
